# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vanilla Laravel Task API",
        "version": "0.1"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "summary": "Get cars paginator",
                "operationId": "1fd4f1fdc0da258c5084bcdce4bbe693",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "summary": "Create car",
                "operationId": "6f519ab2c9767d4d3b1be7cf8f1287ed",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/cars/{carId}": {
            "get": {
                "summary": "Get car",
                "operationId": "ecbf3b7daa28c298fa91175dce46b55f",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "summary": "Update car",
                "operationId": "84659b88dc3842005ebcb038d658796b",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete car",
                "operationId": "27c314ac3f3c855a37572ba03370103d",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "Get users paginator",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "summary": "Create user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/users/{userId}": {
            "get": {
                "summary": "Get user",
                "operationId": "2fc03ba6d92e0712966a5ce4a7d9eaeb",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "summary": "Update user",
                "operationId": "2af19474a22fcb71ca0c134728c7a07a",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "car_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete user",
                "operationId": "76f92b9a1870022f8b5c1c8a899c8a44",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    }
}